backend/requirements.txt
# ====================================
# Flask Framework
# ====================================
Flask==3.0.0
Flask-CORS==4.0.0

# ====================================
# PubNub
# ====================================
pubnub>=7.4.0
pycryptodome>=3.19.0

# ====================================
# Configuration
# ====================================
python-dotenv>=1.0.0

# ====================================
# Database
# ====================================
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# ====================================
# Utilities
# ====================================
python-dateutil>=2.8.2
requests>=2.31.0

# ====================================
# Development
# ====================================
pytest>=7.4.3
pytest-flask>=1.3.0

# ====================================
# Production WSGI Server
# ====================================
gunicorn==21.2.0